add diacard popup & confirm dialog
